astrobase_cli/pyproject.toml
[tool.poetry]
name = "astrobase-cli"
version = "0.1.0"
description = "Astrobase CLI Client"
authors = ["Anthony Corletti <anthony@astrobase.co>"]
readme = "README.md"

[tool.poetry.scripts]
astrobase = "main:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.3.2"}
docker = "^4.4.4"
PyYAML = "^5.4.1"
kubernetes = "^12.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
